Move AddMarkerAtKeysoundPlayPosition to utils.cpp
