+++ sayaslicer/sayaslicer.hpp
@@ -51,5 +51,3 @@
 #endif
 
 const int mainSnaps[] = { 1, 2, 3, 4, 6, 8, 12, 16, 24, 32, 48, 64, 96, 128, 192 };
-
-inline void AddMarkerAtKeysoundPlayPosition(SoundBuffer& buffer, SlicerSettings& settings, bool jumpToMarker);
